+++ src/app.fs
@@ -1,6 +1,0 @@
-﻿module App
-
-open Fable.React
-
-let App() =
-    div [] [ str "Hello from Fable-React Application!"]