+++ webpack.config.js
@@ -1,14 +1,26 @@
 const path = require('path');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
+const StylelintPlugin = require('stylelint-webpack-plugin');
 
 module.exports = {
     entry: './src/app.fsproj',
+    devServer: {
+        contentBase: './dist',
+        open: 'google-chrome',
+        historyApiFallback: true
+    },
     plugins: [
         new CleanWebpackPlugin(),
         new HtmlWebpackPlugin({
             title: 'Theater',
             template: './src/index.html'
+        }),
+        new StylelintPlugin({
+            configFile: 'stylelint.config.js',
+            context: './src',
+            files: '**/*.css',
+            fix: true
         })
     ],
     output: {
@@ -19,6 +31,13 @@ module.exports = {
         rules: [{
             test: /\.fs(x|proj)?$/,
             use: 'fable-loader'
+        }, {
+            test: /\.css$/,
+            use: [
+                'style-loader',
+                'css-loader',
+                'postcss-loader'
+            ]
         }]
     }
 };