+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # 1.0.0
 
 ### Features
-* First working version;
+* **core**: first working version;
+* **core**: integrate React Router;
+
+### Build
+* **core**: add CSS import support;
+* **core**: configure GitHub actions;
+* **core**: update React to v16.12;
+* **core**: enable stylelint;
+* **core**: enable postcss autoprefixer;
+* **core**: add webpack development server;